"GreylistEntry" class is specific to the "simple"
So it's been moved there
