+++ foghornd/__init__.py
@@ -3,9 +3,7 @@
 __all__ = [
     'Foghorn',
     'FoghornSettings',
-    'GreylistEntry',
 ]
 
 from foghorn import Foghorn
-from greylist_entry import GreylistEntry
 from settings import FoghornSettings
